Move perform to its own source file and replace the unbounded overload with ..T15 (max args on the GS side)
